+++ templates/base/packages/nextjs/utils/scaffold-stark/contract.ts
@@ -205,15 +205,21 @@ export type ExtractAbiFunctionNamesScaffold<
 > = ExtractAbiFunctionsScaffold<TAbi, TAbiStateMutability>["name"];
 
 // helper function will only take from interfaces : //TODO: see if we can make it more generic
+// helper function will only take from interfaces : //TODO: see if we can make it more generic
 export type ExtractAbiFunctionsScaffold<
   TAbi extends Abi,
   TAbiStateMutability extends AbiStateMutability = AbiStateMutability,
 > = Extract<
   ExtractAbiInterfaces<TAbi>["items"][number],
-  {
-    type: "function";
-    state_mutability: TAbiStateMutability;
-  }
+  | {
+      state_mutability: TAbiStateMutability;
+    }
+  | Extract<
+      TAbi[number],
+      {
+        type: "function";
+      }
+    >
 >;
 
 // helper function will only take from interfaces : //TODO: see if we can make it more generic
@@ -242,8 +248,6 @@ export type ExtractAbiFunctionScaffold<
     name: TFunctionName;
   }
 >;
-
-// let emerson = singleFunction extends listOfFunctions ? true : false;
 
 export type UseScaffoldArgsParam<
   TContractName extends ContractName,
@@ -290,31 +294,6 @@ export type AbiFunctionOutputs<
   TAbi extends Abi,
   TFunctionName extends string,
 > = ExtractAbiFunctionScaffold<TAbi, TFunctionName>["outputs"];
-
-/*export type AbiEventInputs<TAbi extends Abi, TEventName extends ExtractAbiEventNames<TAbi>> = ExtractAbiEvent<
-  TAbi,
-  TEventName
->["inputs"];
-
-type IndexedEventInputs<
-  TContractName extends ContractName,
-  TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
-> = Extract<AbiEventInputs<ContractAbi<TContractName>, TEventName>[number], { indexed: true }>;
-
-export type EventFilters<
-  TContractName extends ContractName,
-  TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
-> = IsContractDeclarationMissing<
-  any,
-  IndexedEventInputs<TContractName, TEventName> extends never
-    ? never
-    : {
-      [Key in IsContractDeclarationMissing<
-        any,
-        IndexedEventInputs<TContractName, TEventName>["name"]
-      >]?: AbiParameterToPrimitiveType<Extract<IndexedEventInputs<TContractName, TEventName>, { name: Key }>>;
-    }
->;*/
 
 export type UseScaffoldEventHistoryConfig<
   TContractName extends ContractName,